# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EasyTravel API",
        "description": "Documentation for EasyTravel project.",
        "contact": {
            "email": "21732184@live.uem.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/airports/coordinates": {
            "get": {
                "tags": [
                    "airports"
                ],
                "summary": "Get all airports coordinates with URL",
                "description": "Returns all the airports coordinates and basic info.",
                "operationId": "getAirportsCoordinates",
                "responses": {
                    "200": {
                        "description": "Ok.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "description": "All coordinates",
                                            "type": "array",
                                            "items": {
                                                "type": "json"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "airport_id": 216,
                                            "airport_name": "Cootamundra Airport",
                                            "airport_country": "Australia",
                                            "airport_city": "",
                                            "airport_lon": 148.028,
                                            "airport_lat": -34.623901
                                        },
                                        {
                                            "airport_id": 241,
                                            "airport_name": "Adelaide International Airport",
                                            "airport_country": "Australia",
                                            "airport_city": "Adelaide",
                                            "airport_lon": 138.53101,
                                            "airport_lat": -34.945
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Server message that contains the error.",
                                            "type": "string"
                                        },
                                        "exception": {
                                            "description": "Generated exception.",
                                            "type": "string"
                                        },
                                        "file": {
                                            "description": "File that throw the exception.",
                                            "type": "string"
                                        },
                                        "line": {
                                            "description": "Line that throws the exception.",
                                            "type": "integer"
                                        },
                                        "trace": {
                                            "description": "Trace route objects.",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "messagge": "The command failed.",
                                        "exception": "",
                                        "file": "",
                                        "line": 150,
                                        "trace": {
                                            "file": "",
                                            "line": 1,
                                            "content": ""
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/airports/airports": {
            "get": {
                "tags": [
                    "airports"
                ],
                "summary": "Get all airports with URL",
                "description": "Returns all the airports that have an URL.",
                "operationId": "getAirports",
                "responses": {
                    "200": {
                        "description": "Ok.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "description": "All models",
                                            "type": "array",
                                            "items": {
                                                "type": "json"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": [
                                        {
                                            "id": 2286,
                                            "name": "Agen-La Garenne Airport"
                                        },
                                        {
                                            "id": 2289,
                                            "name": "Ajaccio-Napoléon Bonaparte Airport"
                                        },
                                        {
                                            "id": 2306,
                                            "name": "Aurillac Airport"
                                        },
                                        {
                                            "id": 2312,
                                            "name": "Bastia-Poretta Airport"
                                        },
                                        {
                                            "id": 2314,
                                            "name": "Paris Beauvais Tillé Airport"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Server message that contains the error.",
                                            "type": "string"
                                        },
                                        "exception": {
                                            "description": "Generated exception.",
                                            "type": "string"
                                        },
                                        "file": {
                                            "description": "File that throw the exception.",
                                            "type": "string"
                                        },
                                        "line": {
                                            "description": "Line that throws the exception.",
                                            "type": "integer"
                                        },
                                        "trace": {
                                            "description": "Trace route objects.",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "messagge": "The command failed.",
                                        "exception": "",
                                        "file": "",
                                        "line": 150,
                                        "trace": {
                                            "file": "",
                                            "line": 1,
                                            "content": ""
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}